requirements.txt
aiogram==3.4.1
aiosqlite==0.19.0
sqlalchemy[asyncio]==2.0.25
alembic==1.13.1
python-dotenv==1.0.0